todo/requirements.txt
# common
uwsgi==2.0.18
redis==3.3.5
celery==4.3.0
flower==0.9.3
psycopg2-binary==2.8.3
pydevd-pycharm~=192.5728.105

# django
django==2.2.3
django-codemirror-widget==0.5.0
django-celery-beat==1.5.0
django-fernet-fields==0.6
django-cors-headers==3.0.2
django-celery-email==2.0.2
django-model-utils==3.2.0
django-debug-toolbar==2.0
django-taggit==1.1.0
django-filter==2.2.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
djangorestframework-filters==v0.11.1  #FIXME: check if this works for django==2.* and django-filter==2.*?

# tests
pytest==5.0.1
coverage==4.5.4
pytest-django==3.5.1
pytest-cov==2.7.1
factory-boy==2.12.0
